app/requirements.txt
click==8.1.3
Flask==2.2.2
Flask-JWT-Extended==4.4.4
flask-marshmallow==0.14.0
Flask-SQLAlchemy==3.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
marshmallow==3.19.0
packaging==23.0
psycopg2-binary==2.9.5
PyJWT==2.6.0
python-dotenv==0.21.0
six==1.16.0
SQLAlchemy==1.4.46
Werkzeug==2.2.2
